Fix #1972 카운터 모듈에서 날짜 선택 기능을 이용한 XSS 보안 취약점 문제 고침 (XEVE-16-004)

+++ modules/counter/counter.admin.view.php
@@ -28,7 +28,7 @@ class counterAdminView extends counter
 	function dispCounterAdminIndex()
 	{
 		// set today's if no date is given
-		$selected_date = Context::get('selected_date');
+		$selected_date = (int)Context::get('selected_date');
 
 		if(!$selected_date)
 		{
